+++ NEWS.md
@@ -3,6 +3,7 @@
 ### moewishart 0.2 (2025-12-02) (GitHub Only)
 
 * Add EM algorithms (with looking for suitable initial values)
+* Add option `marginal.z` in function `woewishart()`
 
 ### moewishart 0.1 (2025-11-24) (GitHub Only)